backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.0
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
sqlalchemy==2.0.35
alembic==1.13.3
